industrial_automation/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pyserial==3.5
pydantic==2.11.7
sqlite4==0.1.1
reportlab==4.4.3